+++ Mobile_Client/.expo/packager-info.json
@@ -1,10 +1,10 @@
 {
-  "expoServerPort": null,
+  "expoServerPort": 19000,
   "webpackServerPort": null,
-  "expoServerNgrokUrl": null,
-  "packagerNgrokUrl": null,
-  "ngrokPid": null,
-  "devToolsPort": 19003,
+  "expoServerNgrokUrl": "https://pr-hvm.jehfkemsy.mobile-client.exp.direct",
+  "packagerNgrokUrl": "https://packager.pr-hvm.jehfkemsy.mobile-client.exp.direct",
+  "ngrokPid": 16060,
+  "devToolsPort": 19002,
   "packagerPort": null,
   "packagerPid": null
 }
