+++ Mobile_Client/.expo/settings.json
@@ -3,5 +3,5 @@
   "lanType": "ip",
   "dev": true,
   "minify": false,
-  "urlRandomness": "7h-x7s"
+  "urlRandomness": "pr-hvm"
 }
