+++ Mobile_Client/src/auth/landing/style.js
